Move components to src/_components
