+++ apps/kdx/src/app/_components/apps/calendar/recurrence-picker.tsx
@@ -30,8 +30,8 @@ import {
   Toggle,
 } from "@kdx/ui";
 
-import { DatePicker } from "~/components/date-picker";
-import { FrequencyToTxt } from "~/components/frequency-picker";
+import { DatePicker } from "~/app/components/date-picker";
+import { FrequencyToTxt } from "~/app/components/frequency-picker";
 import { tzOffsetText } from "~/helpers";
 
 const freqs = [RRule.DAILY, RRule.WEEKLY, RRule.MONTHLY, RRule.YEARLY];
