+++ apps/kdx/src/app/_components/footer/footer.tsx
