+++ apps/kdx/src/app/_components/header-footer-remover.tsx
