+++ apps/kdx/src/app/apps/page.tsx
@@ -4,8 +4,8 @@ import { auth } from "@kdx/auth";
 import { prisma } from "@kdx/db";
 import { H1, Lead } from "@kdx/ui";
 
-import { KodixApp } from "~/components/app/kodix-app";
-import MaxWidthWrapper from "~/components/max-width-wrapper";
+import { KodixApp } from "~/app/components/app/kodix-app";
+import MaxWidthWrapper from "~/app/components/max-width-wrapper";
 
 export default async function Apps() {
   const session = await auth();
