+++ apps/kdx/src/app/apps/todo/page.tsx
@@ -2,10 +2,10 @@
 
 import { H1, Separator } from "@kdx/ui";
 
-import { columns } from "~/components/apps/todo/columns";
-import { CreateTaskDialogButton } from "~/components/apps/todo/create-task-dialog-button";
-import { DataTable } from "~/components/apps/todo/data-table";
-import MaxWidthWrapper from "~/components/max-width-wrapper";
+import { columns } from "~/app/components/apps/todo/columns";
+import { CreateTaskDialogButton } from "~/app/components/apps/todo/create-task-dialog-button";
+import { DataTable } from "~/app/components/apps/todo/data-table";
+import MaxWidthWrapper from "~/app/components/max-width-wrapper";
 import { api } from "~/trpc/react";
 
 export default function Todo() {
