+++ apps/kdx/src/app/marketplace/page.tsx
@@ -1,7 +1,7 @@
 import { H1, Lead } from "@kdx/ui";
 
-import { KodixApp } from "~/components/app/kodix-app";
-import MaxWidthWrapper from "~/components/max-width-wrapper";
+import { KodixApp } from "~/app/components/app/kodix-app";
+import MaxWidthWrapper from "~/app/components/max-width-wrapper";
 import { api } from "~/trpc/server";
 
 export default async function Apps() {
